+++ crates/forge/tests/utils/running_tests.rs
@@ -51,16 +51,21 @@ pub fn run_test_case(
         load_test_artifacts(&test.path().unwrap().join("target/dev"), package).unwrap();
 
     let ui = Arc::new(UI::default());
+    let cache_dir = Utf8PathBuf::from_path_buf(tempdir().unwrap().keep())
+        .unwrap()
+        .join(DEFAULT_CACHE_DIR);
     rt.block_on(async {
         let target_handles = raw_test_targets
             .into_iter()
             .map(|t| {
+                let cache_dir = cache_dir.clone();
                 tokio::task::spawn_blocking(move || {
                     prepare_test_target(
                         t,
                         &tracked_resource,
                         &NameFilter::All,
                         &PartitionConfig::default(),
+                        &cache_dir,
                     )
                 })
             })
@@ -88,9 +93,7 @@ pub fn run_test_case(
                         fuzzer_seed: 12345,
                         max_n_steps: None,
                         is_vm_trace_needed: false,
-                        cache_dir: Utf8PathBuf::from_path_buf(tempdir().unwrap().keep())
-                            .unwrap()
-                            .join(DEFAULT_CACHE_DIR),
+                        cache_dir: cache_dir.clone(),
                         contracts_data: ContractsData::try_from(
                             test.contracts(&ui).unwrap(),
                             cfg!(feature = "cairo-native"),